refactored appbar
