+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^5.11.8",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
+    "firebase": "^8.2.6",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-qr-scanner": "0.0.11",
